# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Consultation app",
        "title": "IT Services",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Логин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Вход в аккаунт",
                "parameters": [
                    {
                        "description": "login info",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/app.loginReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.loginResp"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Выход из аккаунта",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход из аккаунта",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/registration": {
            "post": {
                "description": "Регистрация",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "Информация о пользователе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.Users"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/app.registerResp"
                        }
                    }
                }
            }
        },
        "/delete-MtM/{idDye}/colorant/{idColorant}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удалить краситель из заявки",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorant-Dye"
                ],
                "summary": "Удалить краситель из заявки",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID красителя",
                        "name": "idColorant",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID заявки",
                        "name": "idDye",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Краситель был удален из заявки",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/list_of_colorants": {
            "get": {
                "description": "Вывести список красителей",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorants"
                ],
                "summary": "Вывести список красителей",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                }
            }
        },
        "/new_colorant": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создать краситель",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorants"
                ],
                "summary": "Создать краситель",
                "parameters": [
                    {
                        "description": "Информация о красителе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                }
            }
        },
        "/delete-service/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Удалить краситель",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorants"
                ],
                "summary": "Удалить краситель",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID красителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                }
            }
        },
        "/update_colorants/{id}": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Обновить краситель",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorants"
                ],
                "summary": "Обновить краситель",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID красителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Информация о красителе",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                }
            }
        },
...
        "/colorant/{id}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Добавить краситель в заявку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Colorants"
                ],
                "summary": "Добавить краситель в заявку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID красителя",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.ColorantsAndOtheres"
                        }
                    }
                }
            }
        },
...
        "/list_of_dyes": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Вывести список заявок",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dyes"
                ],
                "summary": "Вывести список заявок",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Dyes"
                        }
                    },
                    "400": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.Dyes"
                        }
                    },
                    "404": {
                        "description": "Некорректный запрос",
                        "schema": {
                            "$ref": "#/definitions/ds.Dyes"
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/ds.Dyes"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "app.loginReq": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "app.loginResp": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "token_type": {
                    "type": "string"
                }
            }
        },
        "app.registerResp": {
            "type": "object",
            "properties": {
                "ok": {
                    "type": "boolean"
                }
            }
        },
        "ds.ColorantsAndOtheres": {
            "type": "object",
            "properties": {
                "ID_Colorant": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "image": {
                    "type": "string"
                },
                "Description": {
                    "type": "string"
                },
                "Properties": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                }
            }
        },
...
        "ds.Users": {
            "type": "object",
            "properties": {
                "Email_Address": {
                    "type": "string"
                },
                "ID_User": {
                    "type": "integer"
                },
                "Name": {
                    "type": "string"
                },
                "Login": {
                    "type": "string"
                },
                "Password": {
                    "type": "string"
                },
                "Phone": {
                    "type": "string"
                },
                "Role": {
                    "$ref": "#/definitions/role.Role"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}